+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## [Unreleased] 2019-06-18
+## [Unreleased]
 
+### Added
+
+- [TD-1594] Support for proxy login
 - [TD-1893] Use CI_JOB_ID instead of CI_PIPELINE_ID
 
 ## [2.21.0] 2019-06-10